ref: move github testutils to fixtures
