+++ tests/sentry/integrations/github/test_webhooks.py
@@ -4,16 +4,15 @@ from uuid import uuid4
 
 from django.utils import timezone
 
-from sentry import options
-from sentry.models import Commit, CommitAuthor, GroupLink, Integration, PullRequest, Repository
-from sentry.testutils import APITestCase
-
-from .testutils import (
+from fixtures.github import (
     PULL_REQUEST_CLOSED_EVENT_EXAMPLE,
     PULL_REQUEST_EDITED_EVENT_EXAMPLE,
     PULL_REQUEST_OPENED_EVENT_EXAMPLE,
     PUSH_EVENT_EXAMPLE_INSTALLATION,
 )
+from sentry import options
+from sentry.models import Commit, CommitAuthor, GroupLink, Integration, PullRequest, Repository
+from sentry.testutils import APITestCase
 
 
 class WebhookTest(APITestCase):
